requirements.txt
astropy==5.1
asttokens==2.0.8
backcall==0.2.0
bidict==0.22.1
biosppy==1.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.0
filelock==3.9.0
fonttools==4.37.4
fsspec==2023.4.0
future==0.18.2
h5py==3.9.0
heartpy==1.2.7
hrv-analysis==1.0.4
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.0
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.0
nolds==0.5.2
numpy==1.23.3
opencv-python==4.8.0.76
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
pyerfa==2.0.0.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.1
seaborn==0.12.2
shortuuid==1.0.11
six==1.16.0
soundfile==0.11.0
stack-data==0.5.1
sympy==1.12
threadpoolctl==3.1.0
torch==2.1.0+cu118
tornado==6.2
traitlets==5.4.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
wfdb==4.0.0
